Conflicts:
	Tychaia.ProceduralGeneration/AlgorithmChancedBool.cs

+++ Tychaia.ProceduralGeneration.Tests/AlgorithmIncrementWaterDistanceTest.cs
@@ -1,25 +1,20 @@
-//
+// 
 // This source code is licensed in accordance with the licensing outlined
 // on the main Tychaia website (www.tychaia.com).  Changes to the
 // license on the website apply retroactively.
-//
-using System;
+// 
 using Xunit;
 
 namespace Tychaia.ProceduralGeneration.Tests
 {
-
     public class AlgorithmIncrementWaterDistanceTest
     {
-        [Fact, TestFor(typeof(AlgorithmIncrementWaterDistance))]
+        [Fact, TestFor(typeof (AlgorithmIncrementWaterDistance))]
         public void TestNegativeDoubling()
         {
             var inputA = new AlgorithmDebuggingInitialValueDelegate
             {
-                GetValueForPosition = (x, y, z, i, j, k) =>
-                {
-                    return (int)(-1);
-                }
+                GetValueForPosition = (x, y, z, i, j, k) => { return -1; }
             };
 
             int computations;
@@ -37,19 +32,16 @@ namespace Tychaia.ProceduralGeneration.Tests
                     for (var k = 0; k < 16; k++)
                     {
                         Assert.True(
-                            result[i + j * 16 + k * 16 * 16] == -2);
+                            result[i + j*16 + k*16*16] == -2);
                     }
         }
 
-        [Fact, TestFor(typeof(AlgorithmIncrementWaterDistance))]
+        [Fact, TestFor(typeof (AlgorithmIncrementWaterDistance))]
         public void TestPositiveDoubling()
         {
             var inputA = new AlgorithmDebuggingInitialValueDelegate
             {
-                GetValueForPosition = (x, y, z, i, j, k) =>
-                {
-                    return (int)(1);
-                }
+                GetValueForPosition = (x, y, z, i, j, k) => { return 1; }
             };
 
             int computations;
@@ -65,11 +57,10 @@ namespace Tychaia.ProceduralGeneration.Tests
             for (var i = 0; i < 16; i++)
                 for (var j = 0; j < 16; j++)
                     for (var k = 0; k < 16; k++)
-                {
-                    Assert.True(
-                        result[i + j * 16 + k * 16 * 16] == 2);
-                }
+                    {
+                        Assert.True(
+                            result[i + j*16 + k*16*16] == 2);
+                    }
         }
     }
 }
-
